Make auto loading frameworks configurable
Assemblies can be configured with their framework dependencies which the runtime can process automatically when the assembly is loaded. This is the default behavior since most applications will need to load the frameworks for referenced assemblies and the runtime has the information to do it. In some embedded environments assemblies may be loaded into the AppDomain which aren't actually referenced, and in such cases it can be useful to explicitly load required frameworks manually.